Implement memory pool for fixed size objects
Since for reading and writing the page we would be using the
fixed memory block of size 4096 bytes, hence using that through
pool would be highly optimal.
Memory pool unit has been implemented that has low memory foot print
and there is 0 overhead of initialization. It does allocation and free
in O(1).